requirements.txt
pandas~=2.2.3
Flask~=2.2.5
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
spacy~=3.8.4
pypdf~=5.3.0
requests~=2.32.3
gunicorn==23.0.0
SQLAlchemy~=1.4.54
apache-airflow~=2.10.5